M05/tests/requirements.txt
requests==2.31.0
PyJWT==2.8.0
pymysql==1.1.0